refactor(core): split agent/persistence.rs into focused submodules
Split the 3537-line agent/persistence.rs (single impl<C: Channel> Agent<C>
block) into persistence/{mod,history,store,extract,tests}.rs:

- history.rs: load_history with post-load agent mutations
- store.rs: persist_message plus summarization and MemCoT scheduling
- extract.rs: background graph/persona/trajectory/reasoning extraction

The four extraction entrypoints invoked from persist_message are now
pub(super) since the shared impl block spans multiple files. No public API
change: load_history (pub) and persist_message (pub(crate)) keep their
visibility. Tests move verbatim to persistence/tests.rs.

Closes #4878
